// generated/api/admin/models/aPIResponseSuccess.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * BDUI API
 * Спецификация OpenAPI для Backend-Driven UI  приложения

 * OpenAPI spec version: 1.0.0
 */
import type { ApiRepresentation } from './apiRepresentation';

export interface APIResponseSuccess {
  type?: 'success';
  api: ApiRepresentation;
}